feat: Update controller response type to ResponseUtil

+++ backend/app/Http/Controllers/DiscussionController.php
@@ -3,6 +3,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\ResponseUtil;
 use App\Models\Discussion;
 use App\Models\User;
 use App\Service\NotificationService;
@@ -19,46 +20,45 @@ class DiscussionController extends Controller
         try {
             $user = $request->attributes->get('user');
 
-            $discussion = Discussion::findOrFail($id);
+            $discussion = Discussion::find($id);
+
+            if (!$discussion) {
+                return ResponseUtil::NotFound('Discussion not found');
+            }
 
             if ($discussion->owner_id != $user->id) {
                 return response()->json([
                     'success' => false,
                     'message' => 'No permission',
                 ], 403);
+                return ResponseUtil::Unauthorized('No permission');
             }
 
             $discussion->delete();
-            return response()->json([
-                'success' => true,
-                'message' => 'Delete discussion Success',
-            ], 200);
+            return ResponseUtil::Success('Delete discussion success');
         } catch (Exception  $exception) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot delete discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot delete discussion!', $exception->getMessage());
         }
     }
 
-    public function likeDiscussion(Request $request, $discussionId)
+    public function likeDiscussion(Request $request, $id)
     {
         try {
             $user = $request->attributes->get('user');
-            $discussion = Discussion::find($discussionId);
+
+            $discussion = Discussion::find($id);
+
+            if (!$discussion) {
+                return ResponseUtil::NotFound('Discussion not found');
+            }
 
             // comment owner ( receiver_id )
             // book owner ( sender_id )
-            NotificationService::storeDiscussionNotification($user->id, $discussion->owner_id, $discussionId, 'like your discussion!');
+            NotificationService::storeDiscussionNotification($user->id, $discussion->owner_id, $id, 'like your discussion!');
 
             return ReactionService::likeEntity($discussion, $discussion->id, $user->id, 'discussion');
         } catch (Exception  $exception) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot like discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot like discussion!', $exception->getMessage());
         }
     }
 
@@ -69,19 +69,20 @@ class DiscussionController extends Controller
 
             $discussion = Discussion::find($discussionId);
 
+            if (!$discussion) {
+                return ResponseUtil::NotFound('Discussion not found');
+            }
+
             // comment owner ( receiver_id )
             // book owner ( sender_id )
             NotificationService::storeDiscussionNotification($user->id, $discussion->owner_id, $discussionId, 'dislike your discussion!');
 
             return ReactionService::dislikeEntity($discussion, $discussion->id, $user->id, 'discussion');
         } catch (Exception  $exception) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot dislike discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot dislike discussion!', $exception->getMessage());
         }
     }
+
 
 
     public function fetchDiscussionById(Request $request, $id)
@@ -94,19 +95,15 @@ class DiscussionController extends Controller
             if ($request->attributes->has('user')) {
                 $userId = $request->attributes->get('user')->id;
             }
-            $discussion = Discussion::findOrFail($id);
 
-            return response()->json([
-                'success' => true,
-                'message' => 'Successfully get discussion',
-                'data' => $discussion->getData($userId)
-            ], 200);
+            $discussion = Discussion::find($id);
+            if (!$discussion) {
+                return ResponseUtil::NotFound('Discussion not found');
+            }
+
+            return ResponseUtil::Success('Get discussion success', $discussion->getData($userId));
         } catch (Exception $exception) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot get discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot get discussion!', $exception->getMessage());
         }
     }
 
@@ -145,18 +142,9 @@ class DiscussionController extends Controller
                 $items[] = $discussion->getData($userId, true); // userId = null => delete_able = false
             }
 
-            return response()->json([
-                'success' => true,
-                'message' => 'Successfully get discussions',
-                'data' => $items
-            ], 200);
+            return ResponseUtil::Success('Successfully get discussions', $items);
         } catch (Exception $exception) {
-
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot get discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot get discussion!', $exception->getMessage());
         }
     }
 
@@ -174,18 +162,9 @@ class DiscussionController extends Controller
                 $items[] = $discussion->getData($user->id, true);
             }
 
-            return response()->json([
-                'success' => true,
-                'message' => 'Successfully get your discussions',
-                'data' => $items
-            ], 200);
+            return ResponseUtil::Success('Successfully get your discussions', $items);
         } catch (Exception $exception) {
-
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot get your discussions!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot get your discussion!', $exception->getMessage());
         }
     }
 
@@ -197,10 +176,7 @@ class DiscussionController extends Controller
             $discussion = Discussion::findOrFail($id);
 
             if ($discussion->owner_id != $user->id) {
-                return response()->json([
-                    'success' => false,
-                    'message' => 'No permission',
-                ], 403);
+                return ResponseUtil::NoPermission();
             }
 
             $validatedData = $request->validate([
@@ -212,17 +188,9 @@ class DiscussionController extends Controller
             $discussion->image = $validatedData['image'];
             $discussion->save();
 
-            return response()->json([
-                'success' => true,
-                'message' => 'Successfully edited discussion',
-                'data' => $discussion->getData($user->id)
-            ], 200);
+            return ResponseUtil::Success('Successfully edited discussion', $discussion->getData($user->id));
         } catch (Exception $exception) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot edit discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot edit discussion!', $exception->getMessage());
         }
     }
 
@@ -246,17 +214,10 @@ class DiscussionController extends Controller
                 'dislike' => 0,
             ]);
 
-            return response()->json([
-                'success' => true,
-                'message' => 'Successfully added a discussion',
-                'data' => $discussion->getData(),
-            ], 200);
+
+            return ResponseUtil::Success('Successfully added a discussion', $discussion->getData());
         } catch (Exception  $exception) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Cannot add discussion!',
-                'error' => $exception->getMessage()
-            ], 500);
+            return ResponseUtil::ServerError('Cannot add discussion!', $exception->getMessage());
         }
     }
 }
